requirements.txt
binance-futures-connector>=3.6.0
python-dotenv>=1.0.0
requests>=2.25.1